packages/core/pyproject.toml
[project]
name = "dj-ai-studio-core"
version = "0.1.0"
description = "Core library for DJ AI Studio"
requires-python = ">=3.12"
license = { text = "MIT" }
dependencies = [
    "pydantic>=2.0",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.19",
    "greenlet>=3.0",
    "alembic>=1.13",
]

[project.optional-dependencies]
yandex = [
    "yandex-music>=2.2.0",
    "pycryptodome>=3.20",
    "strenum>=0.4",
    "mutagen>=1.47",
]
audio = [
    "librosa>=0.10",
    "numpy>=1.26",
    "scipy>=1.12",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dj_ai_studio"]
